refactor: 메인 뷰컨트롤러에서 메인 뷰 분리
