+++ phonebook/View/MainView/MainView.swift
@@ -1,0 +1,68 @@
+//
+//  MainView.swift
+//  phonebook
+//
+//  Created by eden on 12/11/24.
+//
+
+import UIKit
+import SnapKit
+
+class MainView: UIView {
+    // MARK: - UI Components
+    let titleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "친구 목록"
+        label.font = UIFont.boldSystemFont(ofSize: 24)
+        label.textAlignment = .left
+        return label
+    }()
+    
+    let addButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle("추가", for: .normal)
+        button.titleLabel?.font = UIFont.systemFont(ofSize: 18, weight: .medium)
+        button.setTitleColor(.gray, for: .normal)
+        //        button.addTarget(self, action: #selector(addButtonTapped), for: .touchUpInside)
+        return button
+    }()
+    
+    let tableView: UITableView = {
+        let tableView = UITableView()
+        tableView.rowHeight = 80
+        tableView.backgroundColor = .white
+        return tableView
+    }()
+    
+    // MARK: - Initializer
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupUI()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - Setup Methods
+    private func setupUI() {
+        self.addSubview(titleLabel)
+        self.addSubview(addButton)
+        self.addSubview(tableView)
+        
+        titleLabel.snp.makeConstraints { make in
+            make.top.equalTo(self.safeAreaLayoutGuide.snp.top).offset(0)
+            make.centerX.equalToSuperview()
+        }
+        
+        addButton.snp.makeConstraints { make in
+            make.centerY.equalTo(titleLabel)
+            make.trailing.equalToSuperview().offset(-16)
+        }
+        
+        tableView.snp.makeConstraints { make in
+            make.top.equalTo(titleLabel.snp.bottom).offset(16)
+            make.leading.trailing.bottom.equalToSuperview()
+        }
+    }
+}
